+++ apps/api/src/modules/campaign/campaign.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { CampaignService } from './campaign.service';
+import { PrismaModule } from '../../common/prisma/prisma.module';
+import { QueueModule } from '../../queue/queue.module';
+
+@Module({
+  imports: [PrismaModule, QueueModule],
+  providers: [CampaignService],
+  exports: [CampaignService],
+})
+export class CampaignModule {}